refactor: move message-broker to 1.application
