+++ src/core/4.main/container/container-types.ts
@@ -2,10 +2,10 @@ import { type EmailSender } from '@/communication/1.application/email/email-send
 import { type TemplateCompiler } from '@/core/1.application/compilers/template-compiler'
 import { type Encrypter } from '@/core/1.application/cryptography/encrypter'
 import { type Hasher } from '@/core/1.application/cryptography/hasher'
+import { type MessageBroker } from '@/core/1.application/events/message-broker'
 import { type Translator } from '@/core/1.application/i18n/translator'
 import { type WebApp } from '@/core/3.infra/api/app/web-app'
 import { type SchemaValidator } from '@/core/3.infra/api/validators/schema-validator'
-import { type MessageBroker } from '@/core/3.infra/events/message-broker'
 import { type MongodbClient } from '@/core/3.infra/persistence/mongodb/client/mongodb-client'
 import { type PersistenceClient } from '@/core/3.infra/persistence/persistence-client'
 import { type PostgresClient } from '@/core/3.infra/persistence/postgres/client/postgres-client'
