+++ src/core/4.main/container/events/make-rabbitmq.ts
@@ -1,5 +1,5 @@
 import { getVar, getIntVar } from '@/core/0.domain/utils/var'
-import { type MessageBroker } from '@/core/3.infra/events/message-broker'
+import { type MessageBroker } from '@/core/1.application/events/message-broker'
 import { RabbitmqAdapter } from '@/core/3.infra/events/rabbitmq/rabbitmq-adapter'
 import { logging } from '@/core/4.main/container/logging'
