+++ src/core/4.main/setup/events/index.ts
@@ -1,4 +1,4 @@
-import { type MessageBroker } from '@/core/3.infra/events/message-broker'
+import { type MessageBroker } from '@/core/1.application/events/message-broker'
 import { queues } from '@/core/4.main/setup/events/queues'
 
 export const setupMessageBroker = async (messageBroker: MessageBroker): Promise<void> => {
